+++ apps/twitter-clone/package.json
@@ -2,8 +2,7 @@
   "name": "@apps/twitter-clone",
   "private": true,
   "type": "module",
-  "devDependencies": {
-  },
+  "devDependencies": {},
   "dependencies": {
     "drizzle-orm": "^0.40.1"
   }
