+++ packages/sizuku/src/generator/valibot/generator/relation-valibot-code.ts
@@ -1,5 +1,4 @@
-import { schemaName } from '../../../shared/utils/index.js'
-import { inferInput } from './infer-input.js'
+import { inferInput } from '../../../utils/index.js'
 
 /**
  * Generates Valibot relation schema code from a relation schema AST extraction.
@@ -13,8 +12,8 @@ export function relationValibotCode(
   withType: boolean,
 ): string {
   const base = schema.baseName
-  const relName = schemaName(schema.name)
-  const baseSchema = schemaName(base)
+  const relName = `${schema.name}Schema`
+  const baseSchema = `${base}Schema`
   const fields = schema.fields.map((f) => `${f.name}:${f.definition}`).join(',')
   const obj = `\nexport const ${relName} = v.object({...${baseSchema}.entries,${fields}})`
   if (withType) return `${obj}\n\n${inferInput(schema.name)}\n`
