+++ packages/sizuku/src/generator/zod/generator/relation-zod-code.ts
@@ -1,5 +1,5 @@
 import { schemaName } from '../../../shared/utils/index.js'
-import { infer } from './infer.js'
+import { infer } from '../../../utils/index.js'
 
 /**
  * Generates Zod relation schema code from a relation schema AST extraction.
@@ -13,8 +13,8 @@ export function relationZodCode(
   withType: boolean,
 ): string {
   const base = schema.baseName
-  const relName = schemaName(schema.name)
-  const baseSchema = schemaName(base)
+  const relName = `${schema.name}Schema`
+  const baseSchema = `${base}Schema`
   const fields = schema.fields.map((f) => `${f.name}:${f.definition}`).join(',')
   const obj = `\nexport const ${relName} = z.object({...${baseSchema}.shape,${fields}})`
   if (withType) return `${obj}\n\n${infer(schema.name)}\n`
